requirements.txt
# requirements.txt

flask
stripe
python-dotenv
